ExceptionHandler and Facet Identification Updates; Utilities Added (not complete)
ExceptionHandler:
The exception handler now writes to the GeometryException logger, the
ConcreteASTException logger or the defualt logger.  More loggers will be
added as needed; please let me know if anybody needs a logger to be
handled.

Facet Identification:
Most of the classes added are actually background classes from other
packages that the FacetID relies upon.
Utilites is mostly commented out except a few static methods I used.
GroundedClause is completely translated except one method which relies
on something else that I don't want to look for right now.
The rest of the files are shells of the classes that still need to be
completed. Further updates will come this week.
Primitive, Figure,