Equations and Simplification Translation
So here are all of the Geometry Equations and the Simplification file
translated.  I have not been able to test this because the code is
currently broken, at least 90% because the files that these reference
have not been completed.  Once all of the files have been translated and
uploaded, I should be able to take them and test these out.

If you notice any problems, feel free to change them, but please let me
know what you are doing to the file before uploading it.