Operations and Other Patches
Modified the Simplication class and Equations classes to fit Java
practices, and added the Operations classes.  The only fix that I have
not made is replacing my clunky method of comparison with the keyword
"instanceof", but I will update that and upload that change tomorrow
with my other classes.

My plan for the next couple of days is to get as many errors resolved so
that I can begin writing tests for this.

Let me know if you have any questions or notice any issues.