Tests, Logger Update, ASTException fix
Re-added my ASTException, necessary components were removed. (please ask
next time)

Updated the Logger and LoggerFactory with initialize() and close().  All
that needs to be done to log now is to call initialize() at the
beginning of main and close() at the end of main.

I added a backendTest folder where I put all of my tests to try to
conform to the way Dr. Alvin wants our tests as best as possible.