Complex Circle Generator Test Update
The complex circle generator is almost done; I just have to tweak the
code for the complex triangle generator to build the planar graph.

I also fixed the spacing in the toString in Point because it was
bothering me.  Also the change in the toString broke some of my
assertions in one of my tests...

I plan on completing the complex circle generator and test for it
tomorrow.