More Theorems and Minor Updates on some Definitions
I forgot to change some of the annotations on some of the definitions
and theorems that I previously translated.  So I went through all of the
files that I have recently translated and ensured that they were all
correct.

I added the following theorems:
SumAnglesInTriangle
RelationsOfCongruentAnglesAreCongruent

Theorems that were added but are not completed:
PerpendicularParallelTransversal