QHG can now have edges
I previously did not understand how the quick access worked, so the
QuickAccess class is not needed.

I moved FactConnector because I decided it makes more sense to be with
FactComputer rather than PreComputer.

Made clauseID a public set because it needs to be set within QHG, as I
highly doubt it is being set anywhere else.

Lastly, I added the functions to QHG to add hyperedges from deductions.